Boolean query parameters are now lowercase
The `ConvertToString` method serializes boolean values to `True` and `False`.

The OpenAPI 3.0 specification does not specify how boolean values are to be serialized when they are query parameters. However, when using a nodejs server generated by https://editor.swagger.io/, requests with `queryParam=True` and `queryParam=False` fail the built in validation in the server. The validation code is taken from `swagger-tools`, and can be seen here: https://github.com/apigee-127/swagger-tools/blob/master/lib/validators.js#L484

From this, and from doing some googling, it would seem that people around the internet are of the opinion that boolean query parameters should either be serialized as "true" and "false", or alternatively "1" and "0".

I propose that one should use "true" and "false".