Prevent SingleClientFromPathSegmentsOperationNameGenerator from using brackets in operation names
ConvertPathToName checked, if a path segment contains "{", but if the
parameters enclosed in those brackets includes underscores, the
resulting operation name included "}", which is an illegal character in
operation names for most languages.
The new introduced regex eliminates parameters from the path before
creating the operation name. A test has been added to check that it is
working.