Add ability to template model class names
Models/Dto class names were picked from the spec provided when
generating the code. Now the name of each model can be customized with
a template string, e.g. "My{model}IsNicer".

The main reason for me for adding this is for convinience when searching
for types in an IDE where there also exist other model classes with
similar names it is nicer to be able to add a prefix to the generated
classes so that they don't show as the initial suggestions.

This could already be acomplished by setting the `TypeNameGenerator`
property on the settings. However, that is a bit more complicated and
not as accessible as having this property.

Note that if the template string does not contain "{model}" it will be
ignored. That way the generated code doesn't become too messed up if
someone just specified an explicit name (which all models would then be
called, which is likely not what the user wants).