204 response enhancements
New document generator setting to allow for certain response codes to be treated as a nullable response, even if the response type is void.

On the client generation side, we treat all 2xx responses with the same response type as the 200 response as success responses. Expanded that to include responses with the same type but nullable.

During client generation, if there are multiple success responses where some are nullable, we set the whole method return type as nullable.